Norme

+++ src/libft/ft_strncmp.c
@@ -10,7 +10,7 @@ int	ft_strncmp(const char *s1, const char *s2, size_t n)
 	while (n > 0)
 	{
 		if (*c1 != *c2)
-			return ((int)(*c1) - (int)(*c2));
+			return ((int)(*c1 - *c2));
 		if (*c1 == '\0')
 			return (0);
 		c1++;
